Renamed model file
